Programmatic setup!
* cleaned up the setup_autopilot routine a lot. split everything into separate functions so at least the behavior is *somewhat* encapsulated. they're still very messy and this needs some real reconsideration because it's pretty fragile, but hey at least now it can be called as something other than a module from __main__
* Programmatic setup with custom prefs (specified individually as -p PREFNAME=VALUE -p PREF2=VAL2 and -f prefs.json) and scripts.
* allowed passing multiple scripts and prefs
* still need to write tests
