+++ autopilot/setup/__main__.py
@@ -1,0 +1,3 @@
+from autopilot.setup.setup_autopilot import main
+
+main()
